third_party/blink/web_tests/http/tests/xmlhttprequest/workers/xmlhttprequest-response-type-blob-sync-expected.txt

Request a file as the Blob type with synchronous XMLHttpRequest in a worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/xmlhttprequest-response-type-blob-sync.js
PASS [Worker] req.status is 200
PASS [Worker] req.response.size is 4
PASS [Worker] buffer is "PASS"
PASS [Worker] Sync XHR with responseType="blob" succeeded
PASS successfullyParsed is true

TEST COMPLETE
